playing around with image watermarking

+++ src/business.php
@@ -60,33 +60,80 @@ function upload($id, $image, $imageData){
         $errors++;
         echo $errorMsg;
     }
-   /* else {
-        $imageData['ext'] = (string)$fileExt;
-    }*/
+    else {
+        $imageData['ext'] = $fileExt;
+    }
 
-    if ($errors == 0){
+    if ($errors == 0) {
         $db = get_db();
 
         if ($id == null) {
             $insert = $db->images->insertOne($imageData);
             $imgID = (string)$insert->getInsertedId();
         } else {
-            $db->images->replaceOne(['_id' => new ObjectID($id)], $image);
+            $db->images->replaceOne(['_id' => new ObjectID($id)], $imageData);
+            $imgID = $id;
         }
-        $newFileName = $imgID.".".$fileExt;
-        $fileLoc = "upload/".$newFileName;
-        move_uploaded_file($image, $fileLoc);
-        return true;
+
+        $newFileName = $imgID . "." . $fileExt;
+        $fileLoc = "upload/" . $newFileName;
+
+        // Save the original image
+        if (move_uploaded_file($_FILES['file']['tmp_name'], $fileLoc)) {
+            
+            $wImg = watermark($fileLoc, $imageData['wmark'], $fileExt);
+            if ($wImg) {
+                $wFileLoc = "upload/w" . $newFileName;
+                if (rename($wImg, $wFileLoc)) {
+                    return true;
+                } 
+            } 
+        } else {
+            echo "Failed to save image.";
+            return false;
+        }
+    } else {
+        echo $errorMsg;
+        return false;
     }
-    else {
-    echo $errorMsg;
-    return false;
-    } 
+}
+
+function watermark($imagePath, $wmark, $ext)
+{
+    if ($ext == "jpg" || $ext == "jpeg") {
+        $img = imagecreatefromjpeg($imagePath);
+    } else if ($ext == "png") {
+        $img = imagecreatefrompng($imagePath);
+    }
+
+    $color = imagecolorallocatealpha($img, 255, 0, 0);
+    $font = "static/fonts/arial.ttf";
+
+    // Add the watermark text to the image
+    imagettftext($img, 11, 0, 20, 30, $color, $font, $wmark);
+
+    $wImgPath = tempnam(sys_get_temp_dir(), 'wmark_');
+
+    if ($ext == "jpg" || $ext == "jpeg") {
+        imagejpeg($img, $wImgPath);
+    } else if ($ext == "png") {
+        imagepng($img, $wImgPath);
+    }
+
+    imagedestroy($img);
+
+    return $wImgPath;
 }
 
 function delete_image($id)
 {
     $db = get_db();
-    $db->images->deleteOne(['_id' => new ObjectID($id)]);
+    $image = $db->images->findOne(['_id' => new ObjectID($id)]);
 
+    if ($image) {
+        $fileExt = $image['ext'];
+        $db->images->deleteOne(['_id' => new ObjectID($id)]);
+        $toDelete = (string)$id;
+        unlink("upload/" . $toDelete . "." . $fileExt);
+    }
 }
